+++ gsa/src/web/pages/hosts/component.js
@@ -131,8 +131,6 @@ class HostComponent extends React.Component {
   render() {
     const {
       children,
-      onCloned,
-      onCloneError,
       onCreated,
       onCreateError,
       onDeleted,
@@ -151,8 +149,6 @@ class HostComponent extends React.Component {
         name="host"
         onCreated={onCreated}
         onCreateError={onCreateError}
-        onCloned={onCloned}
-        onCloneError={onCloneError}
         onDeleted={onDeleted}
         onDeleteError={onDeleteError}
         onDownloaded={onDownloaded}
@@ -195,8 +191,6 @@ HostComponent.propTypes = {
   entitiesCounts: PropTypes.counts,
   gmp: PropTypes.gmp.isRequired,
   selectionType: PropTypes.string,
-  onCloneError: PropTypes.func,
-  onCloned: PropTypes.func,
   onCreateError: PropTypes.func,
   onCreated: PropTypes.func,
   onDeleteError: PropTypes.func,